added abstract class for corecreateshape to getlayername
