+++ org.mwc.debrief.legacy/src/Debrief/Tools/Palette/CreateLabel.java
@@ -74,13 +74,16 @@
 package Debrief.Tools.Palette;
 
 import Debrief.Wrappers.LabelWrapper;
-import MWC.GUI.*;
+import MWC.GUI.BaseLayer;
+import MWC.GUI.Layer;
+import MWC.GUI.Layers;
+import MWC.GUI.ToolParent;
 import MWC.GUI.Properties.PropertiesPanel;
-import MWC.GUI.Tools.*;
-import MWC.GUI.Tools.PlainTool.BoundsProvider;
-import MWC.GenericData.*;
+import MWC.GUI.Tools.Action;
+import MWC.GenericData.WorldArea;
+import MWC.GenericData.WorldLocation;
 
-public final class CreateLabel extends PlainTool
+public final class CreateLabel extends CoreCreateShape
 {
 
   /////////////////////////////////////////////////////////////
@@ -88,15 +91,8 @@ public final class CreateLabel extends PlainTool
   ////////////////////////////////////////////////////////////
   /** the properties panel
    */
-  private final PropertiesPanel _thePanel;
+  private PropertiesPanel _thePanel;
 
-  /** the layers we are going to drop this shape into
-   */
-  private final Layers _theData;
-
-  /** the chart we are using (since want our 'duff' item to appear in the middle)
-   */
-  private final BoundsProvider _theBounds;
 
   /////////////////////////////////////////////////////////////
   // constructor
@@ -106,26 +102,20 @@ public final class CreateLabel extends PlainTool
    * @param thePanel panel
    */
   public CreateLabel(final ToolParent theParent,
-                     final PropertiesPanel thePanel,
-                     final Layers theData,
-                     final BoundsProvider bounds,
-                     final String theName,
-                     final String theImage)
+      final PropertiesPanel thePanel,
+      final Layers theData,
+      final BoundsProvider bounds,
+      final String theName,
+      final String theImage)
   {
-    super(theParent, theName, theImage);
+    super(theParent, theName, theImage,theData,bounds);
 
     _thePanel = thePanel;
-    _theData = theData;
-    _theBounds = bounds;
   }
 
   /** get the current visible data area
    * 
    */
-  final protected WorldArea getBounds()
-  {
-    return _theBounds.getViewport();
-  }
 
   /////////////////////////////////////////////////////////////
   // member functions
@@ -141,10 +131,11 @@ public final class CreateLabel extends PlainTool
       final WorldLocation centre = wa.getCentreAtSurface();
 
       final LabelWrapper theWrapper = new LabelWrapper("blank label",
-                                                 centre,
-                                                  MWC.GUI.Properties.DebriefColors.ORANGE);
-
-      Layer theLayer = _theData.findLayer("Misc");
+          centre,
+          MWC.GUI.Properties.DebriefColors.ORANGE);
+      
+      String layerToAddTo = getLayerName();
+      Layer theLayer = _theData.findLayer(layerToAddTo);
       if(theLayer == null)
       {
         theLayer = new BaseLayer();
@@ -153,25 +144,18 @@ public final class CreateLabel extends PlainTool
       }
 
       return new CreateLabelAction(_thePanel,
-                                     theLayer,
-                                     theWrapper,
-                                     _theData);
+          theLayer,
+          theWrapper,
+          _theData);
     }
     else
     {
-        // we haven't got an area, inform the user
-        MWC.GUI.Dialogs.DialogFactory.showMessage("Create Feature",
+      // we haven't got an area, inform the user
+      MWC.GUI.Dialogs.DialogFactory.showMessage("Create Feature",
           "Sorry, we can't create a shape until the area is defined.  Try adding a coastline first");
-          return null;
+      return null;
     }
   }
-
-  /** get the actual instance of the shape we are creating
-   * @return LabelWrapper containing an instance of the new shape
-   * @param centre the current centre of the screen, where the shape should be centred
-   */
- // abstract protected LabelWrapper getShape(WorldLocation centre);
-
 
   ///////////////////////////////////////////////////////
   // store action information
@@ -183,13 +167,13 @@ public final class CreateLabel extends PlainTool
     final PropertiesPanel _thePanel;
     final Layer _theLayer;
     final Debrief.Wrappers.LabelWrapper _theShape;
-		final private Layers _theData;
+    final private Layers _theData;
 
 
     public CreateLabelAction(final PropertiesPanel thePanel,
-                               final Layer theLayer,
-                               final LabelWrapper theShape,
-                               final Layers theData)
+        final Layer theLayer,
+        final LabelWrapper theShape,
+        final Layers theData)
     {
       _thePanel = thePanel;
       _theLayer = theLayer;
@@ -236,9 +220,9 @@ public final class CreateLabel extends PlainTool
       // add the Shape to the layer, and put it
       // in the property editor
       _theLayer.add(_theShape);
-      
+
       if(_thePanel != null)
-      	_thePanel.addEditor(_theShape.getInfo(), _theLayer);
+        _thePanel.addEditor(_theShape.getInfo(), _theLayer);
 
       // and fire the extended event
       _theData.fireExtended();
